Remove Event.ts
